+++ package.h
@@ -10,6 +10,13 @@
 //                                                                 //
 /////////////////////////////////////////////////////////////////////
 
-#define FW_VERSION	"1.1.16"
+#define FW_VERSION	"1.1.17"
+
+#ifndef ARP_TIMEOUT_MS
+#define ARP_TIMEOUT_MS 3000
+#endif
+#ifndef DNS_TIMEOUT_MS
+#define DNS_TIMEOUT_MS 2000
+#endif
 
 #endif