+++ src/app/user/app.h
@@ -10,7 +10,6 @@
 #include "credentials.h"
 #include "./../../../core.h"
 
-
 struct app_settings {
 
   struct fw {
@@ -18,18 +17,26 @@ struct app_settings {
     char         md5[16];
   }fw;
 
+  // user settings below
+  struct sniffer {
+    bool          enabled;
+    uint8_t       channel;
+    uint16_t      loop;
+  }sniffer;
+
 };
 
 // user can edit it
 enum appTopics_ {
-  settings_reset_set_,
+  settings_reset_,
+  sniffer_route_,
   app_not_found
 };
 
-
 // user can edit it
 static const std::map<long, appTopics_> appTopics {
-  { (long)std::hash<std::string>{}("/app/settings/reset/set"),                  settings_reset_set_ },
+  { (long)std::hash<std::string>{}("/app/settings/reset/set"),           settings_reset_},
+  { (long)std::hash<std::string>{}("/app/sniffer"),                      sniffer_route_},
 };
 
 class APP{
@@ -44,6 +51,10 @@ class APP{
     bool getValue(JsonObject& obj, String ref){return false;};
     bool version_changed(){return versionChanged;};
 
+    // user public funcs
+    
+    // user public vars
+
   private:
     // do not delete the following function
     appTopics_ resolveOption(std::map<long, appTopics_> map, String topic);
@@ -52,8 +63,10 @@ class APP{
     bool reset_settings();
     void log_settings();
 
-    uint32_t timeoutInfo;
+    uint32_t timeoutInfo = 5000;
     bool versionChanged = false;
+    uint32_t timeoutSniffer;
+    String msg;
 };
 
 #endif